+++ jkube-kit/enricher/api/src/main/java/org/eclipse/jkube/kit/enricher/api/BaseEnricher.java
@@ -37,7 +37,6 @@ import java.util.Properties;
 
 /**
  * @author roland
- * @since 01/04/16
  */
 public class BaseEnricher implements Enricher {
 
@@ -62,7 +61,7 @@ public class BaseEnricher implements Enricher {
     public BaseEnricher(EnricherContext enricherContext, String name) {
         this.enricherContext = enricherContext;
         // Pick the configuration which is for us
-        this.config = new EnricherConfig(name, enricherContext.getConfiguration());
+        this.config = new EnricherConfig(name, enricherContext);
         this.log = new PrefixedLogger(name, enricherContext.getLog());
         this.name = name;
     }
@@ -94,20 +93,20 @@ public class BaseEnricher implements Enricher {
         return enricherContext.getConfiguration();
     }
 
-    protected String getConfig(Configs.Key key) {
+    protected String getConfig(Configs.Config key) {
         return config.get(key);
     }
 
-    protected boolean hasConfig(Configs.Key key) {
-        return config.get(key) != null;
-    }
-
-    protected String getConfig(Configs.Key key, String defaultVal) {
+    protected String getConfig(Configs.Config key, String defaultVal) {
         return config.get(key, defaultVal);
     }
 
-    protected Map<String, String> getRawConfig() {
-        return config.getRawConfig();
+    protected String getConfigWithFallback(Configs.Config key, String fallbackPropertyKey, String defaultVal) {
+        final String value = getConfig(key, Configs.getFromSystemPropertyWithPropertiesAsFallback(enricherContext.getProperties(), fallbackPropertyKey));
+        if (value != null) {
+            return value;
+        }
+        return defaultVal;
     }
 
     protected EnricherContext getContext() {
@@ -120,7 +119,7 @@ public class BaseEnricher implements Enricher {
      * @return boolean value indicating whether OpenShift or not.
      */
     protected boolean isOpenShiftMode() {
-        Properties properties = getContext().getConfiguration().getProperties();
+        Properties properties = getContext().getProperties();
         if (properties != null) {
             return RuntimeMode.isOpenShiftMode(properties);
         }
@@ -159,7 +158,7 @@ public class BaseEnricher implements Enricher {
      * @param defaultValue default value
      * @return string as image pull policy
      */
-    protected String getImagePullPolicy(ResourceConfig resourceConfig, String defaultValue) {
+    protected static String getImagePullPolicy(ResourceConfig resourceConfig, String defaultValue) {
         if(resourceConfig != null) {
             return resourceConfig.getImagePullPolicy() != null ? resourceConfig.getImagePullPolicy() : defaultValue;
         }
@@ -176,7 +175,7 @@ public class BaseEnricher implements Enricher {
      * @param defaultValue default value
      * @return resolved replica count
      */
-    protected int getReplicaCount(KubernetesListBuilder builder, ResourceConfig xmlResourceConfig, int defaultValue) {
+    protected static int getReplicaCount(KubernetesListBuilder builder, ResourceConfig xmlResourceConfig, int defaultValue) {
         if (xmlResourceConfig != null) {
             List<HasMetadata> items = builder.buildItems();
             for (HasMetadata item : items) {
@@ -190,6 +189,10 @@ public class BaseEnricher implements Enricher {
             return xmlResourceConfig.getReplicas() > 0 ? xmlResourceConfig.getReplicas() : defaultValue;
         }
         return defaultValue;
+    }
+
+    public static String getNamespace(ResourceConfig resourceConfig, String defaultValue) {
+        return Optional.ofNullable(resourceConfig).map(ResourceConfig::getNamespace).orElse(defaultValue);
     }
 
     protected void setProcessingInstruction(String key, List<String> containerNames) {
@@ -225,7 +228,7 @@ public class BaseEnricher implements Enricher {
      */
     protected String getValueFromConfig(String propertyName, String defaultValue) {
         if (getContext().getProperty(propertyName) != null) {
-            return getContext().getProperty(propertyName).toString();
+            return getContext().getProperty(propertyName);
         } else {
             return defaultValue;
         }